backend/requirements.txt
fastapi
uvicorn
qdrant-client
langchain
sentence-transformers
pytest
google-generativeai
python-dotenv